Defensive programming and convert currency

+++ controllers/staff.js
@@ -87,20 +87,20 @@ const psGetRestaurantPromos = new PS({ name: 'get-promos', text: `SELECT pid, po
 const psInsertPromo = new PS({ name: 'get-promo-by-id', text:
 `
 INSERT into Promotions (startDate, endDate, points, percentOff, minSpending, monthsWithNoOrders) values
-($1, $2, $3, $4, $5, $6) RETURNING pid;
+($1, $2, $3, $4, ($5::FLOAT * 100) , $6) RETURNING pid;
 `});
 
-const psInsertRestaurantPromo = new PS({ name: 'insert-rest-promo', text: `INSERT into RestaurantPromos (rid, pid) values ($1, $2);` });
+const psInsertRestaurantPromo = new PS({ name: 'insert-rest-promo', text: `INSERT into RestaurantPromos (rid, pid) values ((SELECT rid FROM Staff WHERE uid = $1), $2);` });
 
 const psEditRestaurantPromos = new PS({ name: 'edit-promo', text:
 `
 UPDATE Promotions
 SET
-startDate = $2, endDate = $3, points = $4, percentOff = $5, minSpending = $6, monthsWithNoOrders  = $7
-WHERE pid = $1;
+startDate = $2, endDate = $3, points = $4, percentOff = $5, minSpending = ($6::FLOAT * 100), monthsWithNoOrders  = $7
+WHERE pid = $1 AND pid in (SELECT pid FROM RestaurantPromos);
 `});
 
-const psDeleteRestaurantPromos = new PS({ name: 'delete-promo', text: 'DELETE FROM Promotions WHERE pid = $1;' });
+const psDeleteRestaurantPromos = new PS({ name: 'delete-promo', text: 'DELETE FROM Promotions WHERE pid = $1 AND (SELECT rid FROM Staff WHERE uid = $2) = (SELECT rid FROM RestaurantPromos WHERE pid = $1);' });
 
 const getStaff = async () => {
     return await db.any(psGetStaff, [uid]);
@@ -137,10 +137,10 @@ const getRestaurantPromos = async (rid) => {
     return await db.any(psGetRestaurantPromos, [rid]);
 }
 
-const insertRestaurantPromos = async (rid, item) => {
+const insertRestaurantPromos = async (uid, item) => {
     const res = await db.tx(async t => {
         const q1 = await t.one(psInsertPromo, [item.startdate, item.enddate, item.points, item.percentoff, item.minspending, item.monthswithnoorders])
-        const q2 = await t.none(psInsertRestaurantPromo, [rid, q1.pid]);
+        const q2 = await t.none(psInsertRestaurantPromo, [uid, q1.pid]);
         return t.batch([q1, q2]);
     });
     return res[0].pid;
@@ -150,8 +150,13 @@ const updateRestaurantPromos = async (item) => {
     await db.none(psEditRestaurantPromos, [item.pid, item.startdate, item.enddate, item.points, item.percentoff, item.minspending, item.monthswithnoorders])
 };
 
-const deleteRestaurantPromos = async (pid) => {
-    await db.none(psDeleteRestaurantPromos, [pid]);
+const deleteRestaurantPromos = async (pid, uid) => {
+    try {
+        await db.none(psDeleteRestaurantPromos, [pid, uid]);
+    } catch (error) {
+        throw 'Unable to delete promotion';
+    }
+    
 };
 
 module.exports = {
